Refactor #26
