+++ src/configs/fontStyle.ts
@@ -11,49 +11,48 @@ export const fontStyles = StyleSheet.create({
     defaultSemibold: {
         fontSize: FontSize.small,
         fontWeight: "600",
-        letterSpacing: 3,
+        letterSpacing: 3
     },
     
     modal: {
-        fontSize: 14,
+        fontSize: FontSize.xSmall,
         fontWeight: "400",
-        letterSpacing: 3,
-        paragraphSpacing: 5,
+        letterSpacing: 3
     },
     
     link: {
         fontSize: FontSize.xSmall,
         fontWeight: "600",
-        letterSpacing: 3,
+        letterSpacing: 3
     },
     
     header: {
         fontSize: FontSize.xxxLarge,
         fontWeight: "700",
-        letterSpacing: 3,
+        letterSpacing: 3
     },
     
     header2: {
         fontSize: FontSize.large,
         fontWeight: "600",
-        letterSpacing: 3,
+        letterSpacing: 3
     },
     
     button: {
         fontSize: FontSize.xLarge,
         fontWeight: "400",
-        letterSpacing: 3,
+        letterSpacing: 3
     },
 
     smallButton: {
         fontSize: FontSize.xSmall,
         fontWeight: "400",
-        letterSpacing: 1,
+        letterSpacing: 1
     },
       
     subtitle: {
         fontSize: FontSize.xxSmall,
         fontWeight: "400",
-        letterSpacing: 3,
+        letterSpacing: 3
     },
 });
