+++ src/constants/buttonsTypes.enum.ts
